Fix CalcInvoiceTotal() Bug
Bug where rounding Hours and DNB hours in Invoice causing totals to be higher than actual invoice calculations.

+++ System/Install/Install.php
@@ -359,8 +359,8 @@ function InArray($keySearch, $array){
 
 							DECLARE Inv_ClientID INT;
 
-							DECLARE TotalHours INT;
-							DECLARE DNBHours INT;
+							DECLARE TotalHours DOUBLE;
+							DECLARE DNBHours DOUBLE;
 
 							DECLARE ExpenseCosts DOUBLE;
